+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 - New: Capes are now downloaded separately from skins. You can choose from NameMC or Minecraft Capes
 - New: Display name setting for Armor Stand
 - Fix: Fixed crash in multiplayer on 1.21.10
-- Fix: Skins from different sources for the same player no longer conflict
-- Other: Support for Armor Poser and Possessive for 1.21.10
+- Fix: Skins from different sources for the same name no longer conflict
+- Fix: Now if there is no image in the source that downloads the skin, the mod will not try to download it
+- Other: Support for Armor Poser for 1.21.10 (Possessive will be later sry)
 
 ||A huge amount of time, nerves, and effort was spent on this update, so I'm going to take a break||